// src/app/open-item-core-data/models/open-item.ts
export class OpenItem {
  id: string|undefined;

  accountingSystemId: string|undefined;

  systemId: string|undefined;
  clientId: string|undefined;
  companyId: string|undefined;
  businessArea: string|undefined;

  fiscalYear: string|undefined;
  documentNumber: string|undefined;
  lineItem: string|undefined;

  accountType: string|undefined;
  account: string|undefined;
  name: string|undefined;
  taxNumber1: string|undefined;
  taxNumber2: string|undefined;
  vatRegistrationNumber: string|undefined;
  customerIdOld: string|undefined;
  buyingGroup: string|undefined;

  documentDate: string|undefined;
  postingDate: string|undefined;
  postingPeriod: string|undefined;

  reference: string|undefined;
  referenceKey: string|undefined;
  ledgerGroup: string|undefined;
  ledger: string|undefined;
  documentCurrency: string|undefined;
  localCurrency: string|undefined;
  rate: string|undefined;
  documentHeaderText: string|undefined;
  documentHeaderReference1: string|undefined;
  documentHeaderReference2: string|undefined;

  documentAmount: string|undefined;
  documentAmountEligibleCashDiscount: string|undefined;
  documentAmountCashDiscount: string|undefined;
  localAmount: string|undefined;
  baselineDate: string|undefined;
  cashDiscountDays1: string|undefined;
  cashDiscountPercent1: string|undefined;
  cashDiscountDays2: string|undefined;
  cashDiscountPercent2: string|undefined;
  cashDiscountDays3: string|undefined;
  cashDiscountPercent3: string|undefined;
  netPaymentDate: string|undefined;
  fixedPaymentTerms: string|undefined;

  paymentBlock: string|undefined;
  taxCode: string|undefined;
  postingKey: string|undefined;
  specialGlIndicator: string|undefined;
  pspElement: string|undefined;
  profitCenter: string|undefined;
  costCenter: string|undefined;

  paymentReference: string|undefined;
  assignment: string|undefined;
  billingDocument: string|undefined;
  reference1: string|undefined;
  reference2: string|undefined;
  reference3: string|undefined;
  orderNumber: string|undefined;
  contract: string|undefined;
  contractType: string|undefined;
  invoiceListNumber: string|undefined;
  branchAccountNumber: string|undefined;
  saleDocumentNumber: string|undefined;
  documentNumberInvoiceReference: string|undefined;
  positionText: string|undefined;

  creationDate: string|undefined;
  creationTime: string|undefined;

  matched: boolean|undefined;
}